# Improved max_concurrency aware ShardManager