--- a/lib/Client.js
+++ b/lib/Client.js
@@ -445,7 +445,7 @@ class Client extends EventEmitter {
             }
 
             for(let i = this.options.firstShardID; i <= this.options.lastShardID; ++i) {
-                this.shards.spawn(i);
+                await this.shards.spawn(i);
             }
         } catch(err) {
             if(!this.options.autoreconnect) {