--- a/lib/gateway/Shard.js
+++ b/lib/gateway/Shard.js
@@ -1854,7 +1854,7 @@ class Shard extends EventEmitter {
                 this.connectTimeout = null;
                 this.status = "ready";
                 this.presence.status = "online";
-                this.client.shards._readyPacketCB();
+                this.client.shards._readyPacketCB(this.id);
 
                 if(packet.t === "RESUMED") {
                     // Can only heartbeat after resume succeeds, discord/discord-api-docs#1619