--- a/lib/gateway/ShardManager.js
+++ b/lib/gateway/ShardManager.js
@@ -146,7 +146,10 @@ class ShardManager extends Collection {
         }
     }
 
-    _readyPacketCB() {
+    _readyPacketCB(shardID) {
+        const rateLimitKey = (shardID % this.concurrency) || 0;
+        this.buckets.set(rateLimitKey, Date.now());
+
         this.tryConnect();
     }