--- a/lib/gateway/ShardManager.js
+++ b/lib/gateway/ShardManager.js
@@ -5,8 +5,6 @@ const Collection = require("../util/Collection");
 const Endpoints = require("../rest/Endpoints");
 const Shard = require("./Shard");
 
-const SHARD_IDENTIFY_RATELIMIT = 5000; // 5s
-
 class ShardManager extends Collection {
     constructor(client) {
         super(Shard);
@@ -117,9 +115,9 @@ class ShardManager extends Collection {
             const rateLimitKey = (shard.id % this.concurrency) || 0;
             const lastConnect = this.buckets.get(rateLimitKey) || 0;
 
-            // has enough time passed since the last connect for this bucket?
+            // has enough time passed since the last connect for this bucket (5s/bucket)?
             // alternatively if we have a sessionID, we can skip this check
-            if(!shard.sessionID && Date.now() - lastConnect < SHARD_IDENTIFY_RATELIMIT) {
+            if(!shard.sessionID && Date.now() - lastConnect < 5000) {
                 continue;
             }