This PR makes ShardManager `max_concurrency` aware. Shards in different `rate_limit_key` (https://discord.com/developers/docs/topics/gateway#sharding-max-concurrency) buckets start in parallel resulting in significantly faster startup times across the board for large scale discord bots. Here is a table (these were tested from intent enabled tokens):

| Old time | New Time | concur | % faster      | shards/total | other              |
|----------|----------|--------|---------------|--------------|--------------------|
| 224.251s | 68.312s  | 16     | ~69.5% (nice) | 16/1024      | getAllUsers: true  |
| 431.849s | 70.237s  | 16     | ~83.7%        | 32/1024      | getAllUsers: false |
| 79.591s  | 76.640s   | 1      | margin of err | 16/16        | getAllUsers: false |